added crop demand
I changed the system! Now the grid is a list of lists where the 0th position is the crop (1=peas, 2=carrots, 3=tomatoes) and the 1st position is the status (0=dirt, 1=growing, 2=harvestable). If we wanted to add extra steps, we could make another status where the player has to water the plant. The list system allows us to tell whether the harvestable crop was the "special_crop" worth 2 points. The timer for crop demand still isn't right, we might have to use the clock.get.ticks method, but I didn't know how to do that so I used basic math.